added encoded byte support for the Appendable builder to speed dot gen
Updated Script builder to support more than 128 scheduled items

+++ src/main/java/com/ociweb/pronghorn/util/math/ScriptedSchedule.java
@@ -5,10 +5,10 @@ import com.ociweb.pronghorn.util.Appendables;
 public class ScriptedSchedule {
 
     public final long commonClock;
-    public final byte[] script;
+    public final int[] script;
     public final int maxRun;
     
-    public ScriptedSchedule(long commonClock, byte[] script, int maxRun) {
+    public ScriptedSchedule(long commonClock, int[] script, int maxRun) {
         this.commonClock = commonClock;
         this.script = script;
         this.maxRun = maxRun;
